+++ mirrord-layer/src/lib.rs
@@ -9,10 +9,7 @@ use std::{
 };
 
 use actix_codec::{AsyncRead, AsyncWrite};
-use common::{
-    CloseFileHook, GetAddrInfoHook, OpenFileHook, OpenRelativeFileHook, ReadFileHook, SeekFileHook,
-    WriteFileHook,
-};
+use common::{GetAddrInfoHook, ResponseChannel};
 use ctor::ctor;
 use envconfig::Envconfig;
 use error::LayerError;
@@ -22,10 +19,8 @@ use futures::{SinkExt, StreamExt};
 use kube::api::Portforwarder;
 use libc::c_int;
 use mirrord_protocol::{
-    ClientCodec, ClientMessage, CloseFileRequest, CloseFileResponse, DaemonMessage, EnvVars,
-    FileRequest, FileResponse, GetAddrInfoRequest, GetAddrInfoResponse, GetEnvVarsRequest,
-    OpenFileRequest, OpenFileResponse, OpenRelativeFileRequest, ReadFileRequest, ReadFileResponse,
-    ResponseError, SeekFileRequest, SeekFileResponse, WriteFileRequest, WriteFileResponse,
+    AddrInfoInternal, ClientCodec, ClientMessage, DaemonMessage, EnvVars, GetAddrInfoRequest,
+    GetEnvVarsRequest,
 };
 use socket::SOCKETS;
 use tcp::TcpHandler;
@@ -33,10 +28,7 @@ use tcp_mirror::TcpMirrorHandler;
 use tokio::{
     runtime::Runtime,
     select,
-    sync::{
-        mpsc::{channel, Receiver, Sender},
-        oneshot,
-    },
+    sync::mpsc::{channel, Receiver, Sender},
     time::{sleep, Duration},
 };
 use tracing::{debug, error, info, trace};
@@ -52,8 +44,7 @@ mod socket;
 mod tcp;
 mod tcp_mirror;
 
-use crate::{common::HookMessage, config::LayerConfig, macros::hook};
-
+use crate::{common::HookMessage, config::LayerConfig, file::FileHandler, macros::hook};
 static RUNTIME: LazyLock<Runtime> = LazyLock::new(|| Runtime::new().unwrap());
 static GUM: LazyLock<Gum> = LazyLock::new(|| unsafe { Gum::obtain() });
 
@@ -89,18 +80,12 @@ fn init() {
     RUNTIME.block_on(start_layer_thread(port_forwarder, receiver, config));
 }
 
-#[allow(clippy::too_many_arguments)]
 async fn handle_hook_message(
     hook_message: HookMessage,
     tcp_mirror_handler: &mut TcpMirrorHandler,
     codec: &mut actix_codec::Framed<impl AsyncRead + AsyncWrite + Unpin + Send, ClientCodec>,
-    // TODO: There is probably a better abstraction for this.
-    open_file_handler: &Mutex<Vec<oneshot::Sender<Result<OpenFileResponse, ResponseError>>>>,
-    read_file_handler: &Mutex<Vec<oneshot::Sender<Result<ReadFileResponse, ResponseError>>>>,
-    seek_file_handler: &Mutex<Vec<oneshot::Sender<Result<SeekFileResponse, ResponseError>>>>,
-    write_file_handler: &Mutex<Vec<oneshot::Sender<Result<WriteFileResponse, ResponseError>>>>,
-    close_file_handler: &Mutex<Vec<oneshot::Sender<Result<CloseFileResponse, ResponseError>>>>,
-    getaddrinfo_handler: &Mutex<Vec<oneshot::Sender<GetAddrInfoResponse>>>,
+    getaddrinfo_handler: &Mutex<Vec<ResponseChannel<Vec<AddrInfoInternal>>>>,
+    file_handler: &mut FileHandler,
 ) {
     match hook_message {
         HookMessage::Tcp(message) => {
@@ -109,143 +94,11 @@ async fn handle_hook_message(
                 .await
                 .unwrap();
         }
-        HookMessage::OpenFileHook(OpenFileHook {
-            path,
-            file_channel_tx,
-            open_options,
-        }) => {
-            debug!(
-                "HookMessage::OpenFileHook path {:#?} | options {:#?}",
-                path, open_options
-            );
-
-            // Lock the file handler and insert a channel that will be used to retrieve the file
-            // data when it comes back from `DaemonMessage::OpenFileResponse`.
-            open_file_handler.lock().unwrap().push(file_channel_tx);
-
-            let open_file_request = OpenFileRequest { path, open_options };
-
-            let request = ClientMessage::FileRequest(FileRequest::Open(open_file_request));
-            let codec_result = codec.send(request).await;
-
-            debug!("HookMessage::OpenFileHook codec_result {:#?}", codec_result);
-        }
-        HookMessage::OpenRelativeFileHook(OpenRelativeFileHook {
-            relative_fd,
-            path,
-            file_channel_tx,
-            open_options,
-        }) => {
-            debug!(
-                "HookMessage::OpenRelativeFileHook fd {:#?} | path {:#?} | options {:#?}",
-                relative_fd, path, open_options
-            );
-
-            open_file_handler.lock().unwrap().push(file_channel_tx);
-
-            let open_relative_file_request = OpenRelativeFileRequest {
-                relative_fd,
-                path,
-                open_options,
-            };
-
-            let request =
-                ClientMessage::FileRequest(FileRequest::OpenRelative(open_relative_file_request));
-            let codec_result = codec.send(request).await;
-
-            debug!("HookMessage::OpenFileHook codec_result {:#?}", codec_result);
-        }
-        HookMessage::ReadFileHook(ReadFileHook {
-            fd,
-            buffer_size,
-            file_channel_tx,
-        }) => {
-            debug!(
-                "HookMessage::ReadFileHook fd {:#?} | buffer_size {:#?}",
-                fd, buffer_size
-            );
-
-            read_file_handler.lock().unwrap().push(file_channel_tx);
-
-            debug!(
-                "HookMessage::ReadFileHook read_file_handler {:#?}",
-                read_file_handler
-            );
-
-            let read_file_request = ReadFileRequest { fd, buffer_size };
-
-            debug!(
-                "HookMessage::ReadFileHook read_file_request {:#?}",
-                read_file_request
-            );
-
-            let request = ClientMessage::FileRequest(FileRequest::Read(read_file_request));
-            let codec_result = codec.send(request).await;
-
-            debug!("HookMessage::ReadFileHook codec_result {:#?}", codec_result);
-        }
-        HookMessage::SeekFileHook(SeekFileHook {
-            fd,
-            seek_from,
-            file_channel_tx,
-        }) => {
-            debug!(
-                "HookMessage::SeekFileHook fd {:#?} | seek_from {:#?}",
-                fd, seek_from
-            );
-
-            seek_file_handler.lock().unwrap().push(file_channel_tx);
-
-            let seek_file_request = SeekFileRequest {
-                fd,
-                seek_from: seek_from.into(),
-            };
-
-            let request = ClientMessage::FileRequest(FileRequest::Seek(seek_file_request));
-            let codec_result = codec.send(request).await;
-
-            debug!("HookMessage::SeekFileHook codec_result {:#?}", codec_result);
-        }
-        HookMessage::WriteFileHook(WriteFileHook {
-            fd,
-            write_bytes,
-            file_channel_tx,
-        }) => {
-            debug!(
-                "HookMessage::WriteFileHook fd {:#?} | length {:#?}",
-                fd,
-                write_bytes.len()
-            );
-
-            write_file_handler.lock().unwrap().push(file_channel_tx);
-
-            let write_file_request = WriteFileRequest { fd, write_bytes };
-
-            let request = ClientMessage::FileRequest(FileRequest::Write(write_file_request));
-            let codec_result = codec.send(request).await;
-
-            debug!(
-                "HookMessage::WriteFileHook codec_result {:#?}",
-                codec_result
-            );
-        }
-        HookMessage::CloseFileHook(CloseFileHook {
-            fd,
-            file_channel_tx,
-        }) => {
-            debug!("HookMessage::CloseFileHook fd {:#?}", fd);
-
-            close_file_handler.lock().unwrap().push(file_channel_tx);
-
-            let close_file_request = CloseFileRequest { fd };
-
-            let request = ClientMessage::FileRequest(FileRequest::Close(close_file_request));
-            let codec_result = codec.send(request).await;
-
-            debug!(
-                "HookMessage::CloseFileHook codec_result {:#?}",
-                codec_result
-            );
+        HookMessage::File(message) => {
+            file_handler
+                .handle_hook_message(message, codec)
+                .await
+                .unwrap();
         }
         HookMessage::GetAddrInfoHook(GetAddrInfoHook {
             node,
@@ -262,93 +115,23 @@ async fn handle_hook_message(
                 service,
                 hints,
             });
-            let codec_result = codec.send(request).await;
 
-            trace!("HookMessage::GetAddrInfo codec_result {:#?}", codec_result);
+            // TODO: Handle this error. We're just ignoring it here and letting -layer crash later.
+            let _codec_result = codec.send(request).await;
         }
     }
 }
 
-#[allow(clippy::too_many_arguments)]
 async fn handle_daemon_message(
     daemon_message: DaemonMessage,
     tcp_mirror_handler: &mut TcpMirrorHandler,
-    // TODO: There is probably a better abstraction for this.
-    open_file_handler: &Mutex<Vec<oneshot::Sender<Result<OpenFileResponse, ResponseError>>>>,
-    read_file_handler: &Mutex<Vec<oneshot::Sender<Result<ReadFileResponse, ResponseError>>>>,
-    seek_file_handler: &Mutex<Vec<oneshot::Sender<Result<SeekFileResponse, ResponseError>>>>,
-    write_file_handler: &Mutex<Vec<oneshot::Sender<Result<WriteFileResponse, ResponseError>>>>,
-    close_file_handler: &Mutex<Vec<oneshot::Sender<Result<CloseFileResponse, ResponseError>>>>,
-    getaddrinfo_handler: &Mutex<Vec<oneshot::Sender<GetAddrInfoResponse>>>,
+    file_handler: &mut FileHandler,
+    getaddrinfo_handler: &Mutex<Vec<ResponseChannel<Vec<AddrInfoInternal>>>>,
     ping: &mut bool,
 ) -> Result<(), LayerError> {
     match daemon_message {
-        DaemonMessage::Tcp(message) => {
-            tcp_mirror_handler.handle_daemon_message(message).await?;
-            Ok(())
-        }
-        DaemonMessage::FileResponse(FileResponse::Open(open_file)) => {
-            debug!("DaemonMessage::OpenFileResponse {open_file:#?}!");
-            debug!("file handler = {:#?}", open_file_handler);
-
-            open_file_handler
-                .lock()?
-                .pop()
-                .ok_or(LayerError::SendErrorFileResponse)?
-                .send(open_file)
-                .map_err(|_| LayerError::SendErrorFileResponse)?;
-
-            Ok(())
-        }
-        DaemonMessage::FileResponse(FileResponse::Read(read_file)) => {
-            // The debug message is too big if we just log it directly.
-            let file_response = read_file
-                .inspect(|success| {
-                    debug!("DaemonMessage::ReadFileResponse {:#?}", success.read_amount)
-                })
-                .inspect_err(|fail| error!("DaemonMessage::ReadFileResponse {:#?}", fail));
-
-            read_file_handler
-                .lock()?
-                .pop()
-                .ok_or(LayerError::SendErrorFileResponse)?
-                .send(file_response)
-                .map_err(|_| LayerError::SendErrorFileResponse)?;
-            Ok(())
-        }
-        DaemonMessage::FileResponse(FileResponse::Seek(seek_file)) => {
-            debug!("DaemonMessage::SeekFileResponse {:#?}!", seek_file);
-
-            seek_file_handler
-                .lock()?
-                .pop()
-                .ok_or(LayerError::SendErrorFileResponse)?
-                .send(seek_file)
-                .map_err(|_| LayerError::SendErrorFileResponse)?;
-            Ok(())
-        }
-        DaemonMessage::FileResponse(FileResponse::Write(write_file)) => {
-            debug!("DaemonMessage::WriteFileResponse {:#?}!", write_file);
-
-            write_file_handler
-                .lock()?
-                .pop()
-                .ok_or(LayerError::SendErrorFileResponse)?
-                .send(write_file)
-                .map_err(|_| LayerError::SendErrorFileResponse)?;
-            Ok(())
-        }
-        DaemonMessage::FileResponse(FileResponse::Close(close_file)) => {
-            debug!("DaemonMessage::CloseFileResponse {:#?}!", close_file);
-
-            close_file_handler
-                .lock()?
-                .pop()
-                .ok_or(LayerError::SendErrorFileResponse)?
-                .send(close_file)
-                .map_err(|_| LayerError::SendErrorFileResponse)?;
-            Ok(())
-        }
+        DaemonMessage::Tcp(message) => tcp_mirror_handler.handle_daemon_message(message).await,
+        DaemonMessage::File(message) => file_handler.handle_daemon_message(message).await,
         DaemonMessage::Pong => {
             if *ping {
                 *ping = false;
@@ -356,6 +139,7 @@ async fn handle_daemon_message(
             } else {
                 Err(LayerError::UnmatchedPong)?;
             }
+
             Ok(())
         }
         DaemonMessage::GetEnvVarsResponse(_) => {
@@ -392,15 +176,11 @@ async fn thread_loop(
 
     // Stores a list of `oneshot`s that communicates with the hook side (send a message from -layer
     // to -agent, and when we receive a message from -agent to -layer).
-    let open_file_handler = Mutex::new(Vec::with_capacity(4));
-    let read_file_handler = Mutex::new(Vec::with_capacity(4));
-    let seek_file_handler = Mutex::new(Vec::with_capacity(4));
-    let write_file_handler = Mutex::new(Vec::with_capacity(4));
-    let close_file_handler = Mutex::new(Vec::with_capacity(4));
     let getaddrinfo_handler = Mutex::new(Vec::with_capacity(4));
 
     let mut ping = false;
     let mut tcp_mirror_handler = TcpMirrorHandler::default();
+    let mut file_handler = FileHandler::default();
 
     loop {
         select! {
@@ -409,12 +189,8 @@ async fn thread_loop(
                     hook_message.unwrap(),
                     &mut tcp_mirror_handler,
                     &mut codec,
-                    &open_file_handler,
-                    &read_file_handler,
-                    &seek_file_handler,
-                    &write_file_handler,
-                    &close_file_handler,
                     &getaddrinfo_handler,
+                    &mut file_handler,
                 ).await;
             }
             daemon_message = codec.next() => {
@@ -422,11 +198,7 @@ async fn thread_loop(
                     if let Err(err) = handle_daemon_message(
                         message,
                         &mut tcp_mirror_handler,
-                        &open_file_handler,
-                        &read_file_handler,
-                        &seek_file_handler,
-                        &write_file_handler,
-                        &close_file_handler,
+                        &mut file_handler,
                         &getaddrinfo_handler,
                         &mut ping,
                     ).await {
@@ -478,17 +250,13 @@ async fn start_layer_thread(
         let env_vars_select = HashSet::from(EnvVars(config.override_env_vars_include));
 
         if !env_vars_filter.is_empty() || !env_vars_select.is_empty() {
-            let codec_result = codec
+            // TODO: Handle this error. We're just ignoring it here and letting -layer crash later.
+            let _codec_result = codec
                 .send(ClientMessage::GetEnvVarsRequest(GetEnvVarsRequest {
                     env_vars_filter,
                     env_vars_select,
                 }))
                 .await;
-
-            debug!(
-                "ClientMessage::GetEnvVarsRequest codec_result {:#?}",
-                codec_result
-            );
 
             let msg = codec.next().await;
             if let Some(Ok(DaemonMessage::GetEnvVarsResponse(Ok(remote_env_vars)))) = msg {